flatten getContext
